Refactor(Race.java): 단일 책임 원칙을 준수해라
단익 책임 원칙에 따라 Race.java의 우승자 기능 책임을 WinnerStrategy.java로 분리함.

수정된 사항:
- WinnerStrategy.java를 생성하여 기능 분리
- 위 수정 사항에 맞게 테스트 코드들 수정
