+++ src/main/java/racingcar/strategy/move/MovementStrategy.java
@@ -1,4 +1,4 @@
-package racingcar.strategy;
+package racingcar.strategy.move;
 
 public interface MovementStrategy {
     boolean shouldMove();
