+++ src/main/java/racingcar/strategy/winner/WinnerStrategy.java
@@ -1,0 +1,9 @@
+package racingcar.strategy.winner;
+
+import racingcar.domain.Car;
+
+import java.util.List;
+
+public interface WinnerStrategy {
+    List<Car> determineWinners(List<Car> cars);
+}
